refactor(workflows): one owner for new-workflow sort order
The same ~35-line query — parent condition for workflows and folders, two
parallel min(sortOrder) reads, fold to a min, subtract one, fall back to 0 —
existed three times:

  lib/workflows/utils.ts               inline in createWorkflowRecord
  lib/workflows/orchestration/...      as a file-private nextWorkflowSortOrder
  lib/workflows/persistence/duplicate  inline, inside the duplicate transaction

The first two are character-identical modulo the table alias. The third had
drifted: it omits isNull(workflow.archivedAt), which the other two apply, so a
folder whose lowest-sortOrder workflow is soft-deleted positioned a *duplicate*
differently from a *create*. The folder-side query agrees in all three, which
marks it as a copy-paste slip rather than intent.

Promotes the helper to lib/workflows/sort-order.ts, taking an optional DbOrTx
so the duplicate path can keep reading inside its transaction. Its own module
rather than utils.ts because duplicate.test.ts and workflow-lifecycle.test.ts
both mock '@/lib/workflows/utils' wholesale — from a separate module the real
query still runs under those suites, so their existing sort-order assertions
keep their meaning and needed no edits.

Note the archived-row behavior itself is not unit-testable here: the shared
dbChainMock does not evaluate WHERE predicates. The guarantee is structural —
one query builder instead of three means the predicate can no longer drift.
